GStreamer: restart camerabin pipeline when changing camera mode.
With Gstreamer 0.10, the camera was always restarded, either by us
(for older Gstreamer versions) or by the camerabin.

With GStreamer 1.0, camerabin doesn't restart the pipeline anymore
when changing the camera mode. Caps are not properly re-negotiated
after changing the mode, leading to errors when trying to capture
alternately images and videos.

Work around this by always restarting the pipeline when changing
the camera mode.

Change-Id: I6a5663a17ca2fc006414cbaa0ffca95286f6dba1
Reviewed-by: Andrew den Exter <andrew.den.exter@qinetic.com.au>